% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/occ-exp.R
\name{agenames}
\alias{agenames}
\title{make labels for age groups}
\usage{
agenames(aw, min.age, max.age)
}
\arguments{
\item{min.age}{the first age}

\item{max.age}{the highest age}

\item{width}{the width of each group}
}
\value{
see make.age.groups
}
\description{
helper function; see make.even.age.groups for more
}

