\name{cmc.decode}
\alias{cmc.decode}
\title{decode a cmc value to a year and month}
\usage{
  cmc.decode(cmc)
}
\arguments{
  \item{cmc}{the cmc value to decode}
}
\value{
  a vector whose first entry is the year and whose second
  entry is the month
}
\description{
  Given a calendar month code (CMC) value, return the
  associated year and month.\cr TODO -- add checks to be
  sure the CMC value makes sense
}

