% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dhstools.R
\docType{package}
\name{dhstools}
\alias{dhstools}
\alias{dhstools-package}
\alias{package-dhstools}
\title{Tools for analyzing Demographic and Health Survey data}
\description{
dhstools has methods useful for analyzing data from
the Demographic and Health Surveys (DHS).
}

