% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/occ-exp.r
\name{exp.in.interval}
\alias{exp.in.interval}
\title{compute the amount of exposure in a given interval}
\usage{
\method{exp}{in.interval}(exp.start, exp.end, int.begin, int.end)
}
\arguments{
\item{exp.start}{the start of the exposure window}

\item{exp.end}{the end of the exposure window}

\item{int.begin}{the start of the time period}

\item{int.end}{the end of the time period}
}
\value{
the amount of exposure in the interval
}
\description{
given the boundaries of a time period (in cmc format),
this fn returns the amount of exposure each person
contributes. so, it computes the amount of exposure
in the period\cr
[exp.start, exp.end)\cr
that happens during\cr
[int.begin, int.end)\cr
TODO:
\itemize{
   \item{ add unit tests }
   \item{ explain the two intervals and how they are used
         (ie, improve documentation)}
}
Note that if there is any missingness in the
exp.start or exp.end variables, then this
assumes no exposure.
}

