% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{get.var}
\alias{get.var}
\title{helper fns for dhstools}
\usage{
get.var(survey.data, var, default = NA)
}
\arguments{
\item{survey.data}{the survey dataset}

\item{var}{either NULL, a column name, or a vector of values}

\item{default}{the default value to fill in if the variable
is not found}
}
\value{
a vector of values whose length is the same as the
        number of rows in survey.data; if var is NULL, this has
        the default values
}
\description{
(NB: get.var and get.weights are taken from the
     networkreporting package)
}
\details{
get a variable from a dataframe or vector

this function was written because a few of the estimator functions
need to use weights, and there are several cases to handle:
the user could pass in a column name, a vector of weights, or
nothing (in which case, the weights should default to 1 for each
row in the dataset). for the special case of getting weights, look
at the curried fn get.weights (right below)
}
\keyword{internal}

