% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/occ-exp.r
\name{make.age.groups}
\alias{make.age.groups}
\title{make an age.groups object}
\usage{
make.age.groups(start, widths, names)
}
\arguments{
\item{start}{the first age}

\item{widths}{the widths of the subsequent age groups}

\item{names}{the names of the age groups}
}
\value{
an object (list) with the widths, names, and
number of age groups, as well as a matrix called
template which has the start and end of each age interval.
the intervals in template are close on the left but not
the right; eg, exp.start of 10 and exp.end of 20
means [10, 20) in terms of exact ages
(this is useful later on, for making individual age
 schedules based on, eg, birth dates)
}
\description{
TODO -- would be nice if this could take either widths
or bins / breaks
}

