% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/occ-exp.r
\name{make.time.periods}
\alias{make.time.periods}
\title{make a time.periods object}
\usage{
make.time.periods(start, durations, names)
}
\arguments{
\item{start}{the start of the time of interest}

\item{durations}{the durations of the subsequent time periods}

\item{names}{the names of the time periods}
}
\value{
an object (list) with the widths, names, and
number of time periods
as well as a matrix called
template which has the start and end of each time period.
the intervals in template are closed on the left but not
on the right; that is, start of 1900 and end of 1910
means [1900, 1910) in terms of exact times.
}
\description{
make a time.periods object
}

