% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/occ-exp.r
\name{OLD.compute.occ.exp}
\alias{OLD.compute.occ.exp}
\title{OLD VERSION - tabulate occurrences and exposures}
\usage{
OLD.compute.occ.exp(formula, data, doi = NULL, age.groups, time.periods,
  time.offsets = NULL, start.obs, end.obs, na.action = "na.pass",
  weights = NULL, unique.exp = NULL, exp.scale = 1)
}
\arguments{
\item{formula}{a formula whose LHS is var that has the timing of events
and whose RHS is covariatess (if any)}

\item{data}{the dataset to use}

\item{doi}{if not NULL, the date of the interview for each row. in this case,
the intervals are treated as years before the interview, so that the
actual time period is potentially different for each row. (see the
\code{intervals} argument, above)}

\item{age.groups}{an age.groups object}

\item{time.periods}{a time.periods object}

\item{time.offsets}{if not NULL, then the time.periods are to be interpreted relative to these times (one for each row). useful for computing quantities like "X months before interview", where interview happened at different times for different respondents}

\item{start.obs}{vector of values (one per row of \code{data}) with the starting point
of the observation window for each row, in CMC format}

\item{end.obs}{vector of values (one per row of \code{data}) with the ending point
of the observation window for each row, in CMC format}

\item{na.action}{what action to take on missing values; defaults to \code{na.pass}}

\item{weights}{the weight to apply to occurrences and exposures; defaults to 1}

\item{unique.exp}{tells which variable identifies unique units for computing exposure.
so, if the data is multiple records for units that experienced
multiple events, this allows us to avoid overcounting exposure}

\item{exp.scale}{TODO}
}
\value{
a list containing
\describe{
   \item{occ.exp}{an array of occurrences and exposures}
   \item{time.periods}{the time intervals}
   \item{age.groups}{the ages}
   \item{exp.scale}{the amount by which exposure is scaled}
}
}
\description{
Given a variable indicating when an event happened,
a time window we are interested in, and possibly a set
of covariates, tabulate counts of event occurences and
exposures in the given time interval.\cr
Note that you have to be careful about observations that
don't experience an event, but still count for exposure;
see the example below.
Also note that the current version can only handle
age intervals of equal width and time intervals of
equal width. (That is, the age and time intervals do
not have to be the same as each other, but all of the
age groups have to have one width and all of the time
groups have to have another width.)
}
\details{
TODO
\itemize{
   \item what if event date is missing?
   \item what about things that vary with the event, eg mother's
                 age when child was born?
   \item sometimes, the order of the indices in the array isn't logical;
                 eg, for a variable with levels {1,2,3,6,9}, the results might be
                 ordered (2,1,6,3,9).
   \item make this capable of handling different dates than
        just CMC codes...
}
}
\examples{
## RECODE so that observations w/ no births show up
  ## in the dataset at least once by giving them a first
  ## birth at the (impossible) CMC code of -1. This ensures
  ## that they never contribute a birth, but that they
  ## still count for exposure.
  ##
  ## NB: this is a key step. if we don't do this,
  ## women who haven't had any births are removed
  ## from the dataset, biasing rates upward...

  bdata.coded <- bdata
  bdata.coded$bdate[ is.na(bdata.coded$bdate) &
                    bdata.coded$bnum == "01" ] <- -1
  bdata.coded <- subset(bdata.coded, ! is.na(bdate) )

  ## NO COVARIATES:
  ## now use compute.occ.exp to get counts of
  ## births and exposure between 1980 and 1990
  ## for ages 0 to 60

  ## TODO -- need to write this example using new
  ##         version

  ## WITH COVARIATES:
  ##  use compute.occ.exp to get counts of
  ## births and exposure by 5-year period
  ## between 1970 and 2005,
  ## for 5-year age groups [0,5), ..., [60,65)
  ## by the covariates
  ## urban, highestedlevel, and religion
  ## (NOTE: this is just illustrative. we wouldn't recommend
  ##  substantively interpreting the results of this example.)

  ## TODO -- need to write this example using new
  ##         version
}

