% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remap-variables.r
\name{rename.vars}
\alias{rename.vars}
\title{quick & dirty fn for renaming variable names...}
\usage{
rename.vars(vnames, fromvars, tovars)
}
\arguments{
\item{vnames}{the list of names to remap}

\item{fromvars}{the current variable name}

\item{tovars}{the new variable name (positions in this
vector should correspond to the ones in
fromvars)}
}
\value{
a vector with \code{vnames} renamed according to
        the mapping from \code{fromvars} to \code{tovars}
}
\description{
it may make sense to take this out and just
use the \code{plyr::rename} function
from the \code{plyr} package
}
\examples{
\dontrun{
   colnames(my.data) <- rename.vars(colnames(my.data),
                                    c("old_var1", "old_var2"),
                                    c("new_var1", "new_var2"))
}
}
\seealso{
\code{\link[plyr]{rename}}
}

