% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{window_intersect}
\alias{window_intersect}
\title{window_intersect}
\usage{
window_intersect(a, b)
}
\arguments{
\item{a}{the first window; here, we use all three components
(start, finish, event)}

\item{b}{the second window; here, we only use the start and finish
time, and ignore any events}
}
\value{
a window containing the exposure and events from a that takes place
        during the time given by b; also, if a has an event and it takes
        place during b, it is included
}
\description{
given two intervals, figure out the intersection between the
first and the second

NOTE that the intervals are treated as half-open (start,finish]
so that an event that happens exactly at time b is counted in
(a,b] but not in (b, c]
}
\details{


TODO - should write a more detailed description
}

