\name{yr.to.cmc}
\alias{yr.to.cmc}
\title{convert a given year (and possibly month) to a
CMC value}
\usage{
  yr.to.cmc(year, month = 1)
}
\arguments{
  \item{year}{the year to convert to a CMC}

  \item{month}{the month to convert to a CMC. if not
  specified, this defaults to 1 (January)}
}
\value{
  the CMC value corresponding to that year
}
\description{
  Given a year and, possibly, a month, return the
  appropriate CMC value
}

